+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - button to reload scenarios
 - pinned requirements
 - loading spinner
+- messaging to handle warnings and errors
 
 ### Fixed
 - added redis to requirements